apps/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.9.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
httpx==0.27.2
prisma==0.15.0
PyJWT==2.10.1
slowapi==0.1.9
gunicorn==22.0.0
uvloop==0.21.0; sys_platform != 'win32'
email-validator==2.1.0
pytest==8.3.5
pytest-asyncio==0.25.3
prometheus-fastapi-instrumentator
